+++ capytaine/bodies/dofs.py
@@ -8,6 +8,12 @@ class RigidBodyDofsPlaceholder:
     def __init__(self, rotation_center=None):
         self.rotation_center = rotation_center
 
+    def __str__(self):
+        return "RigidBodyDofsPlaceholder()"
+
+    def _repr_pretty_(self, p, cycle):
+        p.text(self.__str__())
+
 
 def rigid_body_dofs(rotation_center=None):
     return RigidBodyDofsPlaceholder(rotation_center=rotation_center)